python-services/requirements.txt
celery[redis]==5.4.0
redis==5.0.4
supabase==2.4.6
openai==1.30.0
librosa==0.10.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
ffmpeg-python==0.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
